requirements.txt
Flask==3.0.2
google-generativeai==0.4.1
